refactor(main): Rename lib folder to src
