+++ packages/main/bin/mocks-server
@@ -1,9 +1,10 @@
 #!/usr/bin/env node
 "use strict";
 
-require("../lib/start").start();
+require("../src/start").start();
 
 /*
+Copyright 2019-present Javier Brea
 Copyright 2019 XbyOrange
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
