+++ packages/main/package.json
@@ -25,7 +25,7 @@
   },
   "files": [
     "bin",
-    "lib",
+    "src",
     "index.js"
   ],
   "main": "index.js",
