+++ packages/main/test/Libs.mocks.js
@@ -1,4 +1,5 @@
 /*
+Copyright 2019-present Javier Brea
 Copyright 2019 XbyOrange
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
