+++ packages/main/test/start.spec.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 Javier Brea
+Copyright 2019-present Javier Brea
 Copyright 2019 XbyOrange
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
@@ -16,7 +16,7 @@ const InquirerCli = require("@mocks-server/plugin-inquirer-cli");
 
 const CoreMocks = require("./Core.mocks.js");
 
-const { start } = require("../lib/start");
+const { start } = require("../src/start");
 
 describe("start method", () => {
   let sandbox;
